-Moar texture/patch update
Mr. Smiley II is less piracy now.

It will be a pain in the ass, but I have to do the rest of the textures
and patches, that doesn't have a visible line..
If these are done, I can finally make maps without any changes to the
textures or patches