-Transfer SmileyONE to SmileyII
So when I thought that having a seperate Smiley project was a good idea,
it was as good as dead.

I had the idea that having Episode 4 as "exclusive" to SmileyONE would
be great, but that isn't the case.
Now everyone will be forced to download Smiley II to play the "original"
(*VERY LOUD COUGH*) levels