-anti-"piracy"
We place unfinished textures in unfinished folders, so they won't harass us about piracy anymore.
The build file is now updated to ignore unfinished folders, and as a result Doom/id stuff will not be included in the pk3, because it already exists within DOOM2.WAD

The textures will be back soon, but as I said earlier we won't include them because it's unfinished and it is just a waste of MBs.

Unfinished folders:
:new: graphics/unfinished/*.*
:new: sounds/replace/
:new: patches/unfinished/
:new: textures/unfinished/
:new: sprites/gore/unfinished/
:new: sprites/decoration/unfinished/

Added "WIP" sprites, it will be used on Doom2 monsters until they get proper sprites.

:new: re-enabled Emoley; added 2 very WIP weapons